+++ src/components/betting-controls/betting-fields/BettingField.module.scss
@@ -1,4 +1,4 @@
-@import "../../mixins";
+@import "../../../mixins";
 
 .field {
   text-align: center;
