+++ src/components/betting-controls/betting-fields/BettingFields.module.scss
@@ -1,4 +1,4 @@
-@import "../../mixins";
+@import "../../../mixins";
 
 .bettingField {
   width: 100%;
